+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/skin/command/argument/adapter/DefaultSkinNameArgumentTypeAdapter.java
@@ -1,0 +1,46 @@
+package io.github.sakurawald.fuji.module.initializer.skin.command.argument.adapter;
+
+import com.mojang.brigadier.arguments.ArgumentType;
+import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.builder.RequiredArgumentBuilder;
+import com.mojang.brigadier.context.CommandContext;
+import io.github.sakurawald.fuji.core.auxiliary.minecraft.CommandHelper;
+import io.github.sakurawald.fuji.core.command.argument.adapter.abst.BaseArgumentTypeAdapter;
+import io.github.sakurawald.fuji.core.command.argument.structure.Argument;
+import io.github.sakurawald.fuji.module.initializer.skin.command.argument.wrapper.DefaultSkinName;
+import io.github.sakurawald.fuji.module.initializer.skin.service.SkinService;
+import io.github.sakurawald.fuji.module.initializer.skin.structure.SkinDescriptor;
+import java.util.List;
+import net.minecraft.server.command.ServerCommandSource;
+
+public class DefaultSkinNameArgumentTypeAdapter extends BaseArgumentTypeAdapter {
+
+    @Override
+    protected ArgumentType<?> makeArgumentType() {
+        return StringArgumentType.string();
+    }
+
+    @Override
+    protected Object makeArgumentObject(CommandContext<ServerCommandSource> context, Argument argument) {
+        return new DefaultSkinName(StringArgumentType.getString(context, argument.getArgumentName()));
+    }
+
+    @Override
+    public List<Class<?>> getTypeClasses() {
+        return List.of(DefaultSkinName.class);
+    }
+
+    @Override
+    public List<String> getTypeStrings() {
+        return List.of("default-skin-name");
+    }
+
+    @Override
+    public RequiredArgumentBuilder<ServerCommandSource, ?> makeRequiredArgumentBuilder(String argumentName) {
+        return super.makeRequiredArgumentBuilder(argumentName)
+            .suggests(CommandHelper.Suggestion.iterable(() -> SkinService.getDefaultSkinList()
+                .stream()
+                .map(SkinDescriptor::getSkinName)
+                .toList()));
+    }
+}