+++ fabric/src/main/java/io/github/sakurawald/fuji/module/initializer/skin/command/argument/wrapper/DefaultSkinName.java
@@ -1,0 +1,11 @@
+package io.github.sakurawald.fuji.module.initializer.skin.command.argument.wrapper;
+
+
+import io.github.sakurawald.fuji.core.command.argument.wrapper.abst.StringValue;
+
+public class DefaultSkinName extends StringValue {
+
+    public DefaultSkinName(String value) {
+        super(value);
+    }
+}